refactor: JPA 관련 레퍼지토리 패키지 위치 변경
