+++ src/main/java/kitchenpos/menus/infra/JpaMenuGroupRepository.java
@@ -1,6 +1,8 @@
-package kitchenpos.menus.domain;
+package kitchenpos.menus.infra;
 
 import java.util.UUID;
+import kitchenpos.menus.domain.MenuGroup;
+import kitchenpos.menus.domain.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaMenuGroupRepository extends MenuGroupRepository, JpaRepository<MenuGroup, UUID> {
