+++ src/main/java/kitchenpos/menus/infra/JpaMenuProductRepository.java
@@ -1,7 +1,9 @@
-package kitchenpos.menus.domain;
+package kitchenpos.menus.infra;
 
 import java.util.List;
 import java.util.UUID;
+import kitchenpos.menus.domain.MenuProduct;
+import kitchenpos.menus.domain.MenuProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
