+++ src/main/java/kitchenpos/menus/infra/JpaMenuRepository.java
@@ -1,6 +1,8 @@
-package kitchenpos.menus.domain;
+package kitchenpos.menus.infra;
 
 import java.util.UUID;
+import kitchenpos.menus.domain.Menu;
+import kitchenpos.menus.domain.MenuRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaMenuRepository extends MenuRepository, JpaRepository<Menu, UUID> {
