+++ src/main/java/kitchenpos/products/domain/ProductRepository.java
@@ -11,5 +11,4 @@ public interface ProductRepository {
 
     List<Product> findAll();
 
-    List<Product> findAllByIdIn(List<UUID> ids);
 }
