+++ src/main/java/kitchenpos/products/infra/JpaProductRepository.java
@@ -1,6 +1,8 @@
-package kitchenpos.products.domain;
+package kitchenpos.products.infra;
 
 import java.util.UUID;
+import kitchenpos.products.domain.Product;
+import kitchenpos.products.domain.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaProductRepository extends ProductRepository, JpaRepository<Product, UUID> {
