+++ src/test/java/kitchenpos/menus/domain/InMemoryMenuProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.application;
+package kitchenpos.menus.domain;
 
 import java.util.HashMap;
 import java.util.List;
